requirements.txt
pyasn1==0.4.8
pysmb==1.2.7
